Move Eagle-3 metric logic into separate file
