+++ src/speculators/models/eagle3/metrics.py
@@ -1,0 +1,107 @@
+"""Metrics and loss functions for Eagle3 draft model."""
+
+from functools import partial
+
+import torch
+
+from speculators.models.metrics import (
+    compute_accuracy_single_step,
+    exp_loss_decay,
+    kl_div_loss,
+    loss_function,
+)
+
+
+def align_for_step(
+    logits: torch.Tensor,  # shape: [1, total_seq_len, draft_vocab_size]
+    targets: torch.Tensor,  # shape: [1, total_seq_len, draft_vocab_size]
+    loss_mask: torch.Tensor | None,  # shape: [1, total_seq_len]
+    prev_correct: torch.Tensor | None,  # shape: [1, total_seq_len]
+    ttt_step: int,
+):
+    """Align logits, targets, loss_mask, and prev_correct for a given ttt_step.
+
+    There are no target values for the last ttt_step tokens, so we mask them out
+    before computing the loss/accuracy. Likewise, there are no logits for the first
+    ttt_step tokens, so we mask them out.
+    This is equivalent to shifting the target values by ttt_step + 1 to the left
+    which puts them in the correct position for the generated tokens.
+    e.g.
+        indices of targets = [1, 2, 3, 4, 5, 6, 7, 8, 9]
+        indices of logits for ttt_step_0 = [1, 2, 3, 4, 5, 6, 7, 8, 9] # no shift
+        indices of logits for ttt_step_1 = [2, 3, 4, 5, 6, 7, 8, 9, 10] # shift by 1
+        indices of logits for ttt_step_2 = [3, 4, 5, 6, 7, 8, 9, 10, 11] # shift by 2
+    The indices for the loss_mask need to be kept in line with the targets indices
+    """
+    logits = logits[:, :-ttt_step] if ttt_step > 0 else logits
+    # shape: [1, total_seq_len - ttt_step, draft_vocab_size]
+    targets = targets[:, ttt_step:]
+    # shape: [1, total_seq_len - ttt_step, draft_vocab_size]
+    if loss_mask is not None:
+        loss_mask = loss_mask[:, ttt_step:]
+        # shape: [1, total_seq_len - ttt_step]
+    if prev_correct is not None:
+        # Align with draft starts
+        prev_correct = prev_correct[:, :-ttt_step] if ttt_step > 0 else prev_correct
+        # shape: [1, total_seq_len - ttt_step]
+    return logits, targets, loss_mask, prev_correct
+
+
+def compute_metrics(
+    logits: torch.Tensor,
+    targets: torch.Tensor,
+    loss_mask: torch.Tensor | None,
+    prev_correct: torch.Tensor | None,
+    ttt_step: int,
+    ttt_step_loss_decay: float,
+) -> tuple[torch.Tensor, dict]:
+    """Compute metrics for a given ttt_step.
+
+    Args:
+        logits: The logits for the current ttt_step.
+        targets: The targets for the current ttt_step.
+        loss_mask: The loss mask for the current ttt_step.
+        prev_correct: The previous correct predictions for the current ttt_step.
+        ttt_step: The current ttt_step.
+        ttt_step_loss_decay: The loss decay for the current ttt_step.
+
+    Effects:
+        Modifies prev_correct in place.
+
+    Returns:
+        Loss value and metrics dictionary.
+    """
+    s_logits, s_targets, s_loss_mask, s_prev_correct = align_for_step(
+        logits, targets, loss_mask, prev_correct, ttt_step
+    )
+
+    seq_len = s_logits.shape[1]
+    if s_loss_mask is None:
+        s_loss_mask = torch.ones(1, seq_len, device=s_logits.device, dtype=torch.bool)
+
+    pos_idx = torch.full(
+        (1, seq_len), ttt_step, device=s_logits.device, dtype=torch.long
+    )
+
+    s_loss = loss_function(
+        s_logits,
+        s_targets,
+        s_loss_mask,
+        pos_idx,
+        loss_fn=kl_div_loss,
+        decay_fn=partial(exp_loss_decay, gamma=ttt_step_loss_decay),
+    )
+
+    pred_ids = torch.argmax(s_logits, dim=-1)
+    target_ids = torch.argmax(s_targets, dim=-1)
+
+    s_full_acc, s_cond_acc = compute_accuracy_single_step(
+        pred_ids, target_ids, s_loss_mask, s_prev_correct
+    )
+
+    s_metrics = {}
+    s_metrics[f"loss_{ttt_step}"] = s_loss.detach().clone()
+    s_metrics[f"full_acc_{ttt_step}"] = s_full_acc
+    s_metrics[f"cond_acc_{ttt_step}"] = s_cond_acc
+
+    return s_loss, s_metrics
